requirements.txt
Django==5.0.7
pillow==10.4.0
numpy==1.26.4
qrcode==7.3.1
opencv-python-headless
pyzbar  
